feat: addd searching service
